pyproject.toml
[project]
name = "todo-list"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "djoser>=2.3.1",
    "python-dotenv>=1.0.1",
]
